Move speed conversion into PropulsionSystem
